fix: reflect refreshed data on user identity page
Using different source of data (user vs userData) with unmatching
setUser caused that values (e.g. firstname) could not be edited and if
they could be, they would not be refreshed by new values.

Using a copy of initially loaded data and refreshing this copy on
refresh makes the editing and refresh possible.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>